knowledgeGraph/pyproject.toml
[project]
name = "knowledgegraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.26",
    "pypdf>=5.6.1",
    "python-dotenv>=1.1.1",
    "pyvis>=0.3.2",
    "streamlit>=1.46.1",
]
